HuddleService now sends an "event" through the Notifier that a member has been registered.
The HuddleServiceMemberRegistrationNotificationTest drives the domain through notifying via email.

NEXT: implement the SendGrid email notifier that sends the actual email. Will need to pull the "add to calendar" code somewhere accessible.

(Note that the HuddleService takes a MemberRepository as a dependency, and I'm not happy about that -- will need a NotifierService instead so it can coordinate between the Huddle and Member repositories.)