Ensemble can now be canceled. It can be in one of three states: Scheduled, Completed, or Canceled.
EnsembleService handles cancel requests from the Adapter, and the Admin dashboard controller has a POST endpoint for cancelling.

Admin Ensemble detail has a big cancel button (still need to show the current status and not show cancel if it's either completed or canceled).

MemberId and EnsembleId have a shorter toString() for logging and error message purposes.